+++ src/Orchestra/Model/User.php
@@ -169,7 +169,7 @@ class User extends Eloquent implements UserInterface, RemindableInterface, Recip
      */
     public function is($roles)
     {
-        $userRoles = $this->resolveRolesAsArray();
+        $userRoles = $this->getRoles();
 
         // For a pre-caution, we should return false in events where user
         // roles not an array.
@@ -196,7 +196,7 @@ class User extends Eloquent implements UserInterface, RemindableInterface, Recip
      */
     public function isAny(array $roles)
     {
-        $userRoles = $this->resolveRolesAsArray();
+        $userRoles = $this->getRoles();
 
         // For a pre-caution, we should return false in events where user
         // roles not an array.
@@ -223,23 +223,7 @@ class User extends Eloquent implements UserInterface, RemindableInterface, Recip
      */
     public function isNot($roles)
     {
-        $userRoles = $this->resolveRolesAsArray();
-
-        // For a pre-caution, we should return false in events where user
-        // roles not an array.
-        if (! is_array($userRoles)) {
-            return false;
-        }
-
-        // We should ensure that any given roles does not match the current
-        // user, consider it as OR condition.
-        foreach ((array) $roles as $role) {
-            if (in_array($role, $userRoles)) {
-                return false;
-            }
-        }
-
-        return true;
+        return ! $this->is($roles);
     }
 
     /**
@@ -250,32 +234,20 @@ class User extends Eloquent implements UserInterface, RemindableInterface, Recip
      */
     public function isNotAny(array $roles)
     {
-        $userRoles = $this->resolveRolesAsArray();
-
-        // For a pre-caution, we should return false in events where user
-        // roles not an array.
-        if (! is_array($userRoles)) {
-            return false;
-        }
-
-        // We should ensure that any given roles does not match the current user,
-        // consider it as OR condition.
-        foreach ($roles as $role) {
-            if (! in_array($role, $userRoles)) {
-                return true;
-            }
-        }
-
-        return false;
+        return ! $this->isAny($roles);
     }
 
     /**
-     * Resolve roles into array for checking
-     *  
+     * Get roles name as an array.
+     *
      * @return array
      */
-    protected function resolveRolesAsArray()
+    public function getRoles()
     {
-        return $this->roles()->lists('name');
+        // If the relationship is already loaded, avoid re-querying the
+        // database and instead fetch the collection.
+        $roles = (array_key_exists('roles', $this->relations) ? $this->relations['roles'] : $this->roles());
+
+        return $roles->lists('name');
     }
 }