+++ tests/UserTest.php
@@ -93,16 +93,14 @@ class UserTest extends \PHPUnit_Framework_TestCase
 
     /**
      * Test Orchestra\Model\User::is() method
-     * 
+     *
      * @test
      */
     public function testIsMethod()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->times(4)->andReturn($relationship);
-        $relationship->shouldReceive('lists')->times(4)->andReturn(array('admin', 'editor'));
+        $model->shouldReceive('getRoles')->times(4)->andReturn(array('admin', 'editor'));
 
         $this->assertTrue($model->is('admin'));
         $this->assertFalse($model->is('user'));
@@ -119,11 +117,9 @@ class UserTest extends \PHPUnit_Framework_TestCase
      */
     public function testIsMethodWhenInvalidRolesIsReturned()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->times(4)->andReturn($relationship);
-        $relationship->shouldReceive('lists')->times(4)->andReturn('foo');
+        $model->shouldReceive('getRoles')->times(4)->andReturn('foo');
 
         $this->assertFalse($model->is('admin'));
         $this->assertFalse($model->is('user'));
@@ -134,16 +130,14 @@ class UserTest extends \PHPUnit_Framework_TestCase
 
     /**
      * Test Orchestra\Model\User::isNot() method
-     * 
+     *
      * @test
      */
     public function testIsNotMethod()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->times(4)->andReturn($relationship);
-        $relationship->shouldReceive('lists')->times(4)->andReturn(array('admin', 'editor'));
+        $model->shouldReceive('getRoles')->times(4)->andReturn(array('admin', 'editor'));
 
         $this->assertTrue($model->isNot('user'));
         $this->assertFalse($model->isNot('admin'));
@@ -160,31 +154,27 @@ class UserTest extends \PHPUnit_Framework_TestCase
      */
     public function testIsNotMethodWhenInvalidRolesIsReturned()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->times(4)->andReturn($relationship);
-        $relationship->shouldReceive('lists')->times(4)->andReturn('foo');
+        $model->shouldReceive('getRoles')->times(4)->andReturn('foo');
 
-        $this->assertFalse($model->isNot('admin'));
-        $this->assertFalse($model->isNot('user'));
+        $this->assertTrue($model->isNot('admin'));
+        $this->assertTrue($model->isNot('user'));
 
-        $this->assertFalse($model->isNot(array('admin', 'editor')));
-        $this->assertFalse($model->isNot(array('admin', 'user')));
+        $this->assertTrue($model->isNot(array('admin', 'editor')));
+        $this->assertTrue($model->isNot(array('admin', 'user')));
     }
 
     /**
      * Test Orchestra\Model\User::isAny() method
-     * 
+     *
      * @test
      */
     public function testIsAnyMethod()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->twice()->andReturn($relationship);
-        $relationship->shouldReceive('lists')->twice()->andReturn(array('admin', 'editor'));
+        $model->shouldReceive('getRoles')->twice()->andReturn(array('admin', 'editor'));
 
         $this->assertTrue($model->isAny(array('admin', 'user')));
         $this->assertFalse($model->isAny(array('superadmin', 'user')));
@@ -198,11 +188,9 @@ class UserTest extends \PHPUnit_Framework_TestCase
      */
     public function testIsAnyMethodWhenInvalidRolesIsReturned()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->twice()->andReturn($relationship);
-        $relationship->shouldReceive('lists')->times(2)->andReturn('foo');
+        $model->shouldReceive('getRoles')->twice()->andReturn('foo');
 
         $this->assertFalse($model->isAny(array('admin', 'editor')));
         $this->assertFalse($model->isAny(array('admin', 'user')));
@@ -210,18 +198,17 @@ class UserTest extends \PHPUnit_Framework_TestCase
 
     /**
      * Test Orchestra\Model\User::isNotAny() method
-     * 
+     *
      * @test
      */
     public function testIsNotAnyMethod()
     {
-        $model = m::mock('\Orchestra\Model\User[roles]');
-        $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
 
-        $model->shouldReceive('roles')->twice()->andReturn($relationship);
-        $relationship->shouldReceive('lists')->twice()->andReturn(array('admin', 'editor'));
+        $model->shouldReceive('getRoles')->times(3)->andReturn(array('admin', 'editor'));
 
-        $this->assertTrue($model->isNotAny(array('admin', 'user')));
+        $this->assertTrue($model->isNotAny(array('administrator', 'user')));
+        $this->assertFalse($model->isNotAny(array('user', 'editor')));
         $this->assertFalse($model->isNotAny(array('admin', 'editor')));
     }
 
@@ -233,14 +220,35 @@ class UserTest extends \PHPUnit_Framework_TestCase
      */
     public function testIsNotAnyMethodWhenInvalidRolesIsReturned()
     {
+        $model = m::mock('\Orchestra\Model\User[getRoles]');
+
+        $model->shouldReceive('getRoles')->twice()->andReturn('foo');
+
+        $this->assertTrue($model->isNotAny(array('admin', 'editor')));
+        $this->assertTrue($model->isNotAny(array('admin', 'user')));
+    }
+
+    /**
+     * Test Orchestra\Model\User::getRoles() method.
+     *
+     * @test
+     */
+    public function testGetRolesMethod()
+    {
         $model = m::mock('\Orchestra\Model\User[roles]');
         $relationship = m::mock('\Illuminate\Database\Eloquent\Relations\BelongsToMany[lists]');
 
-        $model->shouldReceive('roles')->twice()->andReturn($relationship);
-        $relationship->shouldReceive('lists')->twice()->andReturn('foo');
+        $model->shouldReceive('roles')->once()->andReturn($relationship);
+        $relationship->shouldReceive('lists')->once()->andReturn(array('admin', 'editor'));
 
-        $this->assertFalse($model->isNotAny(array('admin', 'editor')));
-        $this->assertFalse($model->isNotAny(array('admin', 'user')));
+        $this->assertEquals(array('admin', 'editor'), $model->getRoles());
+        /*
+
+        $stub = $model->roles();
+
+        $this->assertInstanceOf('\Illuminate\Database\Eloquent\Relations\BelongsToMany', $stub);
+        $this->assertInstanceOf('\Orchestra\Model\Role', $stub->getQuery()->getModel());
+        */
     }
 
     /**